feat: refactor RegisterForm import paths
