+++ src/components/auth/RegisterForm.tsx
@@ -2,7 +2,7 @@ import { Form } from "../common/Form.tsx";
 import { initialUserData } from "../../types/auth.ts";
 import { registerFields } from "../../config/formFields.ts";
 import { Heading } from "../common";
-import Loading from "../common/Loading";
+import Loading from "../common/Loading.tsx";
 import useRegisterForm from "../../hooks/useRegisterForm.ts";
 import styles from "../../styles/components/Register.module.scss";
 import "react-toastify/dist/ReactToastify.css";
