+++ src/pages/Registrationpage.tsx
@@ -1,5 +1,5 @@
 import { PageBanner, UspBanner } from "../components/common";
-import RegisterForm from "../components/register/RegisterForm.tsx";
+import RegisterForm from "../components/auth/RegisterForm.tsx";
 
 const RegisterPage = () => {
   return (
